% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_finding.R
\name{calculate.adjusted.trait}
\alias{calculate.adjusted.trait}
\title{Adjust mean of phenotype distribution to compensate for genetic effect.}
\usage{
calculate.adjusted.trait(trait.characteristic, freq, beta)
}
\arguments{
\item{trait.characteristic}{Numeric mean of phenotype.}

\item{freq}{Numeric allele frequency of variant.}

\item{beta}{Numeric regression coefficient.}
}
\value{
Numeric trait mean with expected genetic effect removed.
}
\description{
The below equation is valid for linear regression only. The original
implementation tried to handle other GLM linkers, but that wasn't
truly supported in the rest of the package, so it is removed for
the time being.
}
