% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mse.R
\name{compute.beta.mse}
\alias{compute.beta.mse}
\title{Compute MSE adjustment for beta correction.}
\usage{
compute.beta.mse(biased.beta, biased.se, debiased.beta)
}
\arguments{
\item{biased.beta}{Numeric input regression coefficient from discovery.}

\item{biased.se}{Numeric input coefficient standard error from discovery.}

\item{debiased.beta}{Numeric MLE adjusted regression coefficient.}
}
\value{
Numeric weighted average coefficient reflecting MSE adjustment.
}
\description{
Computes weighted average of biased and strictly debiased
regression coefficient to report as intermediate result.
}
\details{
Note that this was apparently completely disabled in the original
implementation, such that the output reflected the MLE correction only.
}
