% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mse.R
\name{compute.ci.mse}
\alias{compute.ci.mse}
\title{Compute MSE adjustment for confidence interval correction.}
\usage{
compute.ci.mse(biased.beta, biased.se, existing.ci, p.thresh)
}
\arguments{
\item{biased.beta}{Numeric input regression coefficient from discovery.}

\item{biased.se}{Numeric input coefficient standard error from discovery.}

\item{existing.ci}{Numeric MLE adjusted confidence interval.}
}
\value{
Numeric weighted average confidence interval reflecting MSE adjustment.
}
\description{
Computes weighted average of biased and strictly debiased
regression coefficient confidence interval to report as intermediate result.
}
\details{
Note that this was apparently completely disabled in the original
implementation, such that the output reflected the MLE correction only.
}
