% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication.R
\name{compute.power}
\alias{compute.power}
\title{Compute power to detect a signal given certain assumptions.}
\usage{
compute.power(beta, freq, sample.size, p.threshold)
}
\arguments{
\item{beta}{Numeric additive regression coefficient for a biallelic variant. In the
context of power calculations, this is meant to be a proxy for the true variant effect.}

\item{freq}{Numeric allele frequency for a biallelic variant. This should probably
be the frequency in the population (or sample) you'll be using to try to detect a signal,
and should be as accurate and precise as possible.}

\item{sample.size}{Numeric number of subjects in study sample. This should be as accurate
as possible; so for example, in the context of GWAS meta-analysis, do not simply provide
the \emph{maximum} sample size available in theory, but rather the actual sample size available
for the variant in question.}

\item{p.threshold}{Numeric p-value of significance required to call a variant
as detected. Depending on what phase of a study you're in, this value may differ substantially.
For studies testing millions of variants in discovery, this will likely be 5*10^-8 (genome-wide
significance) or lower; for replication, this may be more like \frac{0.05}{variants brought into replication}.}
}
\value{
Numeric approximate power to detect a signal given the input descriptors of
that signal and study sample.
}
\description{
The power to detect an association is, in short, the probability
of the variant exceeding the required test statistic threshold
as set under the null hypothesis when the alternative hypothesis
is true. In formal terms, this becomes 1 - \chi^2_1((\chi^2_1)^{-1}(1 - \alpha), ncp_v)
where the noncentrality parameter is N\frac{variance explained}{residual variance}.
How fine grained you want to get with this calculation is really up to personal
preference.
}
\details{
Which version of these values you provide to the function depends on the question
you want to ask. In the context of predicting replication, you probably want
to provide the \emph{debiased discovery beta} and the \emph{replication sample size and frequency}.
But different permutations of these values will answer different questions.
}
\examples{
freq <- 0.25
beta <- 0.1
sample.size <- 10000
p.threshold <- 5e-8
compute.power(beta, freq, sample.size, p.threshold)
}
\seealso{
compute.variance.explained
}
