% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication.R
\name{compute.variance.explained}
\alias{compute.variance.explained}
\title{Compute approximate phenotypic variance explained by a genetic variant.}
\usage{
compute.variance.explained(beta, freq)
}
\arguments{
\item{beta}{Numeric additive regression coefficient for a biallelic variant.}

\item{freq}{Numeric allele frequency for a biallelic variant.}
}
\value{
Numeric approximate variance explained for the variant, assuming the
association is small in magnitude.
}
\description{
Given certain very important assumptions, notably that the genetic effect
of a variant is nearly 0, the genetic variance explained under additivity
for biallelic variants is simply 2B^2f(1-f), for B regression coefficient,
f allele frequency.
}
\details{
Note that the genetic variance explained described above breaks down when
the regression coefficient substantially exceeds 0. It is slightly ridiculous
to consider an equation that self-destructs when it actually has something
to estimate, but there you go.
}
\examples{
freq <- 0.25
beta <- 0.05
varexp <- compute.variance.explained(beta, freq)
}
\seealso{
compute.power
}
