% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_winners_curse.R
\name{correct.winners.curse}
\alias{correct.winners.curse}
\title{Correct Winner's Curse in GWAS association data.}
\usage{
correct.winners.curse(
  input.file,
  output.file,
  trait.mean = 0,
  discovery.threshold = 5e-08,
  header = TRUE,
  sep = "\\t"
)
}
\arguments{
\item{input.file}{Character vector name of input data file.}

\item{output.file}{Character vector name of output file
to which results should be reported.}

\item{trait.mean}{Numeric mean of trait used for discovery.
If NA, the function will expect mean values to be specified in
each row of the input data file.}

\item{discovery.threshold}{Numeric threshold on p-value used
in discovery phase to select variants for replication. If NA,
the function will expect threshold values to be specified in
each row of the input data file.}

\item{header}{Logical indicating whether input data file
has a header.}

\item{sep}{Character vector indicating delimiter in input data.}
}
\description{
This function accepts a file containing summary information
describing the discovery association metrics for a set
of significant GWAS variants. The function will attempt
to adjust those values to correct for the apparent
"Winner's Curse" effect on those estimates, which depends
most notably on the p-value threshold used in discovery
to select variants for replication. Both MLE and MSE
corrections are emitted; generally, MLE has performed better
in testing, but that is not guaranteed to be accurate.
}
\details{
This function is a reimplementation of the original C++
program that accompanied the paper at
\url{https://doi.org/10.1371/journal.pgen.1006916}. The implementation
is largely improved and streamlined. As a brief summary of changes:
\itemize{
\item obviously, it's an R package now
\item linear regression is explicitly the only method supported. that
isn't actually a change from the original, as the cooked-in
alternate GLM support was not directly part of the paper
\item the input and output formats are slightly changed, which is fine
because they weren't really document to begin with whoops.
the formats are as follows:
\item Input:
\itemize{
\item beta (discovery)
\item standard error (discovery)
\item sample size (discovery)
\item allele frequency (discovery)
\item phenotype distribution mean in study (optional)
\item p-value threshold to select variant for replication (optional, though
required if phenotype distribution mean is specified as a column)
\item beta (replication)
\item standard error (replication)
\item sample size (replication)
\item allele frequency (replication)
}
\item Output:
\itemize{
\item The same first columns as input. A header is added or slightly adjusted as needed.
If no per-row trait mean/p-value threshold was provided in the input, they will be
appended as per-row values in columns 9 and 10 of the output
\item beta (MLE adjustment for Winner's Curse)
\item 95\% CI on beta, lower bound (MLE adjustment for Winner's Curse)
\item 95\% CI on beta, upper bound (MLE adjustment for Winner's Curse)
\item beta (MSE adjustment for Winner's Curse)
\item 95\% CI on beta, lower bound (MSE adjustment for Winner's Curse)
\item 95\% CI on beta, upper bound (MSE adjustment for Winner's Curse)
}
\item Additional modifications to input and output format are available based
on formal parameters to the function, as described in the parameter documentation.
\item In the original package, standard error of the input regression coefficient was being
reestimated. This was originally intended in advance of support for non-linear GLM,
but that support never manifested, and as such, the correction was actually somewhat
counterproductive, seeing as:
\itemize{
\item there is a simple, fast closed-form solution to standard error
for additive genetic effect under HWE for linear models (so it just went slow for no reason)
\item any deviation from HWE (which is expected for some significant variants) caused
the correction to be inaccurate
\item linear model genetic effect standard error is independent of beta (yes, it's true), and
as such, no correction is actually needed (again, only for linear regression)
In light of the above, the correction method is wholly removed, and will only be added
back in if other GLM support is attempted at some later date.
}
}
}
