% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_finding.R
\name{debias.beta}
\alias{debias.beta}
\title{Compute MLE debiased regression coefficient.}
\usage{
debias.beta(
  beta.biased,
  stderr.biased,
  freq,
  trait.characteristic,
  p.threshold.init
)
}
\arguments{
\item{beta.biased}{Numeric input biased regression coefficient.}

\item{stderr.biased}{Numeric input coefficient standard error.}

\item{freq}{Numeric variant allele frequency.}

\item{trait.characteristic}{Numeric phenotype mean.}

\item{p.threshold.init}{Numeric reported p-value threshold used
to bring variants into replication.}
}
\value{
Numeric maximum likelihood estimate of debiased regression coefficient.
}
\description{
Adjusts input regression coefficient according to Winner's Curse
logic and threshold used for discovery. Most of the logic present
in the original implementation has been removed to improve maintainability
and legibility.
}
\details{
Some of the original implementation was designed to support
non-linear GLMs, but the implementation was never finished and
left in an unstable state. That content has been removed in favor
of making the linear regression content transparent and testable.
The additional content may be added back in at a later date, maybe.

Note that papers often misreport the actual p-value used to select
variants for replication. In some cases, a reasonable approximation
of the true threshold used can be derived from the observed p-values,
but if the investigators just spiked in manually selected variants,
the entire model somewhat breaks down.
}
