% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_finding.R
\name{debiasing.func}
\alias{debiasing.func}
\title{Compute optimization function for iteratively debiasing regression coefficient.}
\usage{
debiasing.func(beta.debiased.init, p.thresh, beta.biased, stderr.biased)
}
\arguments{
\item{beta.debiased.init}{Numeric input attempted debiased value of regression coefficient.}

\item{p.thresh}{Numeric discovery p-value threshold.}

\item{beta.biased}{Numeric input biased regression coefficient.}

\item{stderr.biased}{Numeric input (possibly) biased standard error.}
}
\value{
Numeric optimization function value for beta debiasing routine.
}
\description{
This function was substantially more complicated in the original implementation,
but as with many other functions, this only actually supported linear regression
in practice, so other GLM linker support is removed for the time being.
}
