% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication.R
\name{expected.replication.count}
\alias{expected.replication.count}
\title{For a set of variants with data about both discovery and replication,
determine the number of variants expected to successfully replicate.}
\usage{
expected.replication.count(
  variant.effect,
  study.frequency,
  study.sample.size,
  p.threshold
)
}
\arguments{
\item{variant.effect}{Numeric variant effect size. In theory, this is meant
to be the true underlying effect estimate, though in practice this will likely
be estimated from another dataset.}

\item{study.frequency}{Numeric allele frequency of variant within the study sample.
Deviations of this number from the actual study allele frequency will negatively
impact the replication count estimation. In particular, reference allele frequency
from external datasets is discouraged for use here.}

\item{study.sample.size}{Numeric number of subjects in study sample.
Deviations of this number from the actual study sample size will negatively
impact the replication count estimation. In particular, this should \emph{not}
be \emph{maximum sample size} but rather \emph{actual sample size} per variant.}
}
\value{
Numeric expected number of variants from the combined variant dataset
that are expected to replicate, given the provided study parameters.
}
\description{
This function computes power to detect each variant in replication, and
combines this information across all variants to create an expected value
of number of variants to replicate.
}
\details{
This step is really challenging to interpret correctly. If all values are
correctly specified, the quality (or lack thereof) of this calculation will
be determined by the extent to which (1) the discovery regression coefficient
is biased by, for example, the Winner's Curse, and (2) the extent to which
shared genetic effect is actually present across the discovery and replication
samples.

However, it is in fact most likely that the values will \emph{not} be correctly
specified in this step, in which case the interpretation is much more vague.
Specifically, the sample size and allele frequency data available for most studies
is both inaccurate and imprecise, due to a variety of challenging factors.
Any amount of misspecification of these values will cause problems; but very notably,
the tendency of meta-analyses (or the GWAS catalog) to specify \emph{maximum} sample
size instead of \emph{actual per-variant} sample size causes systematic overestimation
of the power to replicate. There is no easy answer to this problem, and unfortunately
resources like the GWAS catalog are actually moving farther away from any rigorous
standard of data collection in favor of a simplified standardization of input information,
to maximize collection. Thus it seems likely that it will actually become
less possible over time to get systematic replication assessments in the field,
and interested investigators should conduct these calculations within their own
datasets where they have access to improved information.
}
\examples{
beta <- c(0.1, 0.2)
f <- c(0.22, 0.12)
n <- c(10000, 5000)
p.thresh <- c(1e-5, 1e-6)
expected.replication.count(beta, f, n, p.thresh)
}
\seealso{
compute.power, compute.variance.explained
}
