% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stderr_expectation.R
\name{linear.stderr.expectation}
\alias{linear.stderr.expectation}
\title{Compute expected standard error for genetic additive
coefficient in linear regression.}
\usage{
linear.stderr.expectation(freq, n.init)
}
\arguments{
\item{freq}{Numeric allele frequency.}

\item{n.init}{Integer input sample size.}
}
\value{
Numeric predicted standard error for
linear regression additive genetic effect estimate.
}
\description{
The linear regression standard error for the additive
model is independent of the regression coefficient,
and has a simple closed-form solution.
}
\details{
The original implementation of this function had theoretical
support for other GLM linkers, but the rest of the package
did not support that, so I'm not really sure what it was
doing here to begin with.

This equation assumes Hardy Weinberg, which basically
just renders it a worse approximation of the error than
the input error provided by the user. Nevertheless, this
is recorded here for posterity.
}
