% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_binomial_test.R
\name{poisson.binom.test}
\alias{poisson.binom.test}
\title{Performs an approximate test of a simple null hypothesis
about the probability of success in a Poisson binomial
series.}
\usage{
poisson.binom.test(
  x,
  n,
  p = 0.5,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{Numeric number of success counts.}

\item{n}{Numeric number of trials.}

\item{p}{Numeric sequence of probabilities for each trial in turn.}

\item{alternative}{Character, one of "two.sided", "less", or "greater."
Analogous to the same parameter in binom.test.}
}
\value{
Numeric probability of the alternative hypothesis
given the input parameters.
}
\description{
This function extends stats::binom.test for the Poisson
binomial distribution. Most of the original binom.test
functionality is removed in favor of a streamlined
calculation of the p-value itself.
}
\details{
To be entirely frank, I don't have a great deal of confidence
in whether this thing works or not. It seems like a clever trick
to me, but then, it's entirely possible it isn't implemented
in the source package itself for a reason.
}
\examples{
x <- 20
n <- 30
p <- seq(0, 1, length.out = 10)
poisson.binom.test(x, n, p, "greater")
}
