% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_finding.R
\name{calculate.ci}
\alias{calculate.ci}
\title{Compute asymptotic 95\% confidence interval around debiased regression
coefficient estimate.}
\usage{
calculate.ci(beta.debiased, stderr.debiased, p.thresh)
}
\arguments{
\item{beta.debiased}{Numeric parameter estimate.}

\item{stderr.debiased}{Numeric standard error of the mean.}

\item{p.thresh}{Numeric p-value boundary for signed Z statistic.}
}
\value{
Numeric vector corresponding to provided p-value boundary.
}
\description{
This just computes the 95\% confidence interval. The original implementation
had a wildly overcomplicated zero solving method that is entirely
unnecessary given normal theory.
}
