% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_winners_curse.R
\name{correct.winners.curse}
\alias{correct.winners.curse}
\title{Correct Winner's Curse in GWAS association data}
\usage{
correct.winners.curse(
  input.file,
  output.file,
  regression.type = c("linear", "logistic", "poisson"),
  trait.mean = NA,
  stderr.precision = NA,
  discovery.threshold = 5e-08,
  header = TRUE,
  invariant.stderr = FALSE
)
}
\arguments{
\item{input.file}{Character vector name of input data file.}

\item{output.file}{Character vector name of output file
to which results should be reported.}

\item{regression.type}{Character vector name of regression
type used to generate association data.}

\item{trait.mean}{Numeric mean of trait used for discovery.
If NA, the function will expect mean values to be specified in
each row of the input data file.}

\item{stderr.precision}{Numeric requested precision of standard
error calculation.}

\item{discovery.threshold}{Numeric threshold on p-value used
in discovery phase to select variants for replication.}

\item{header}{Logical indicating whether input data file
has a header.}

\item{invariant.stderr}{Logical indicating whether to use
the simplifying assumption that the standard error does not
change along with shrinking effect estimate.}
}
\description{
TBD
}
