% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_winners_curse.R
\name{correct.winners.curse}
\alias{correct.winners.curse}
\title{Correct Winner's Curse in GWAS association data.}
\usage{
correct.winners.curse(
  input.file,
  output.file,
  trait.mean = 0,
  discovery.threshold = 5e-08,
  header = TRUE,
  sep = sep
)
}
\arguments{
\item{input.file}{Character vector name of input data file.}

\item{output.file}{Character vector name of output file
to which results should be reported.}

\item{trait.mean}{Numeric mean of trait used for discovery.
If NA, the function will expect mean values to be specified in
each row of the input data file.}

\item{discovery.threshold}{Numeric threshold on p-value used
in discovery phase to select variants for replication. If NA,
the function will expect threshold values to be specified in
each row of the input data file.}

\item{header}{Logical indicating whether input data file
has a header.}

\item{sep}{Character vector indicating delimiter in input data.}
}
\description{
TBD
}
